# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Config Service",
        "title": "Config Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1323",
    "basePath": "/api",
    "paths": {
...
        "/services/{service_version_id}/features/{feature_version_id}/keys": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get keys for a feature",
                "produces": [
                    "application/json"
                ],
                "summary": "Get keys for a feature",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service version ID",
                        "name": "service_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Feature version ID",
                        "name": "feature_version_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/key.KeyItemDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service version ID",
                        "name": "service_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Feature version ID",
                        "name": "feature_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create key request",
                        "name": "createKeyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/services/{service_version_id}/features/{feature_version_id}/keys/name-taken/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check if key name is taken",
                "produces": [
                    "application/json"
                ],
                "summary": "Check if key name is taken",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service version ID",
                        "name": "service_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Feature version ID",
                        "name": "feature_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BooleanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/services/{service_version_id}/features/{feature_version_id}/keys/{key_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a key",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service version ID",
                        "name": "service_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Feature version ID",
                        "name": "feature_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Key ID",
                        "name": "key_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/key.KeyDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service version ID",
                        "name": "service_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Feature version ID",
                        "name": "feature_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Key ID",
                        "name": "key_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update key request",
                        "name": "updateKeyRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a key",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Service version ID",
                        "name": "service_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Feature version ID",
                        "name": "feature_version_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Key ID",
                        "name": "key_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.ValueTypeKind": {
            "type": "string",
            "enum": [
                "string",
                "integer",
                "decimal",
                "boolean",
                "json"
            ],
            "x-enum-varnames": [
                "ValueTypeKindString",
                "ValueTypeKindInteger",
                "ValueTypeKindDecimal",
                "ValueTypeKindBoolean",
                "ValueTypeKindJson"
            ]
        },
        "db.ValueValidatorType": {
            "type": "string",
            "enum": [
                "required",
                "min_length",
                "max_length",
                "min",
                "max",
                "min_decimal",
                "max_decimal",
                "regex",
                "json_schema",
                "valid_json",
                "valid_integer",
                "valid_decimal",
                "valid_regex"
            ],
            "x-enum-varnames": [
                "ValueValidatorTypeRequired",
                "ValueValidatorTypeMinLength",
                "ValueValidatorTypeMaxLength",
                "ValueValidatorTypeMin",
                "ValueValidatorTypeMax",
                "ValueValidatorTypeMinDecimal",
                "ValueValidatorTypeMaxDecimal",
                "ValueValidatorTypeRegex",
                "ValueValidatorTypeJsonSchema",
                "ValueValidatorTypeValidJson",
                "ValueValidatorTypeValidInteger",
                "ValueValidatorTypeValidDecimal",
                "ValueValidatorTypeValidRegex"
            ]
        },
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "handler.BooleanResponse": {
            "type": "object",
            "required": [
                "value"
            ],
            "properties": {
                "value": {
                    "type": "boolean"
                }
            }
        },
...
        "handler.CreateKeyRequest": {
            "type": "object",
            "required": [
                "name",
                "validators",
                "valueTypeId"
            ],
            "properties": {
                "defaultValue": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "validators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ValidatorRequest"
                    }
                },
                "valueTypeId": {
                    "type": "integer"
                }
            }
        },
        "handler.CreateResponse": {
            "type": "object",
            "required": [
                "newId"
            ],
            "properties": {
                "newId": {
                    "type": "integer"
                }
            }
        },
...
        "handler.UpdateKeyRequest": {
            "type": "object",
            "required": [
                "validators"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "validators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ValidatorRequest"
                    }
                }
            }
        },
...
        "handler.ValidatorRequest": {
            "type": "object",
            "required": [
                "errorText",
                "parameter",
                "validatorType"
            ],
            "properties": {
                "errorText": {
                    "type": "string"
                },
                "parameter": {
                    "type": "string"
                },
                "validatorType": {
                    "$ref": "#/definitions/db.ValueValidatorType"
                }
            }
        },
...
        "key.KeyDto": {
            "type": "object",
            "required": [
                "canEdit",
                "description",
                "id",
                "name",
                "validators",
                "valueType",
                "valueTypeId",
                "valueTypeName"
            ],
            "properties": {
                "canEdit": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "validators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validation.ValidatorDto"
                    }
                },
                "valueType": {
                    "$ref": "#/definitions/db.ValueTypeKind"
                },
                "valueTypeId": {
                    "type": "integer"
                },
                "valueTypeName": {
                    "type": "string"
                }
            }
        },
        "key.KeyItemDto": {
            "type": "object",
            "required": [
                "description",
                "id",
                "name",
                "valueType",
                "valueTypeId",
                "valueTypeName"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "valueType": {
                    "$ref": "#/definitions/db.ValueTypeKind"
                },
                "valueTypeId": {
                    "type": "integer"
                },
                "valueTypeName": {
                    "type": "string"
                }
            }
        },
...
        "validation.ValidatorDto": {
            "type": "object",
            "required": [
                "errorText",
                "isBuiltIn",
                "parameter",
                "validatorType"
            ],
            "properties": {
                "errorText": {
                    "type": "string"
                },
                "isBuiltIn": {
                    "type": "boolean"
                },
                "parameter": {
                    "type": "string"
                },
                "validatorType": {
                    "$ref": "#/definitions/db.ValueValidatorType"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}